backend/pyproject.toml
[tool.poetry]
name = "chat_tts_api"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.115.12"
uvicorn = "0.34.2"
pydantic = "2.11.4"
python-dotenv = "1.1.0"
minio = "7.2.15"
numpy = "1.26.4"
pandas = "2.2.3"
ffmpy = "0.5.0"
aiofiles = "24.1.0"
python-multipart = "0.0.20"
